refactor to bionemo-llm datamodule.py
